# arweave-uploader-main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API server for really arweave chain. Use OpenAPI 2.0 specification.",
        "title": "really arweave API",
        "contact": {
            "name": "BillZong",
            "email": "billzong@163.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/images": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "upload image to arweave chain",
                "operationId": "images-upload",
                "parameters": [
                    {
                        "type": "file",
                        "description": "图片，文件类型目前仅支持png,jpg,jpeg",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e.Response"
                        }
                    }
                }
            }
        },
        "/images/uri": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "give image uri and resave it to arweave chain",
                "operationId": "images-save-uri",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router_v1_images.SaveImageUriReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ccian.cc_really_arweave-api_router_v1_images.SaveImageUriReq": {
            "type": "object",
            "properties": {
                "uri": {
                    "description": "图片URI",
                    "type": "string"
                }
            }
        },
        "e.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "error code of this api server",
                    "type": "integer"
                },
                "data": {
                    "description": "data"
                },
                "field": {
                    "description": "special field prompt, especially when an error occu",
                    "type": "string"
                },
                "message": {
                    "description": "error message",
                    "type": "string"
                }
            }
        },
        "router_v1_images.SaveImageUriReq": {
            "type": "object",
            "properties": {
                "uri": {
                    "description": "图片URI",
                    "type": "string"
                }
            }
        }
    }
}